[feat] Tratamento de exceções Part 1

+++ src/main/java/com/digitalhouse/clinicaodontologicag6/service/impl/DentistaServiceImpl.java
@@ -2,6 +2,9 @@ package com.digitalhouse.clinicaodontologicag6.service.impl;
 
 import com.digitalhouse.clinicaodontologicag6.entity.DentistaEntity;
 import com.digitalhouse.clinicaodontologicag6.entity.dto.DentistaDTO;
+import com.digitalhouse.clinicaodontologicag6.exception.NotFoundException;
+import com.digitalhouse.clinicaodontologicag6.exception.UserAlreadyExistsException;
+import com.digitalhouse.clinicaodontologicag6.exception.VariableNullException;
 import com.digitalhouse.clinicaodontologicag6.repository.IDentistaRepository;
 import com.digitalhouse.clinicaodontologicag6.repository.IPacienteRepository;
 import com.digitalhouse.clinicaodontologicag6.service.IClinicaService;
@@ -28,42 +31,57 @@ public class DentistaServiceImpl implements IClinicaService<DentistaDTO>, UserDe
     BCryptPasswordEncoder passwordEncoder;
 
     @Override
-    public DentistaDTO create(DentistaDTO dentistaDTO) {
-        DentistaEntity dentistaEntity = mapperDTOToEntity(dentistaDTO);
-        String password = passwordEncoder.encode(dentistaEntity.getPassword());
-        dentistaEntity.setPassword(password);
-        dentistaEntity = dentistaRepository.save(dentistaEntity);
-        dentistaDTO = mapperEntityToDTO(dentistaEntity);
+    public DentistaDTO create(DentistaDTO dentistaDTO) throws UserAlreadyExistsException {
+        try {
+            DentistaEntity dentistaEntity = mapperDTOToEntity(dentistaDTO);
+            String password = passwordEncoder.encode(dentistaEntity.getPassword());
+            dentistaEntity.setPassword(password);
+            dentistaEntity = dentistaRepository.save(dentistaEntity);
+            dentistaDTO = mapperEntityToDTO(dentistaEntity);
+        } catch (Exception e) {
+            throw new UserAlreadyExistsException("Usuário já cadastrado");
+        }
         return dentistaDTO;
     }
 
     @Override
-    public DentistaDTO getById(Long id) {
-        DentistaEntity dentista = dentistaRepository.findById(id).get();
+    public DentistaDTO getById(Long id) throws NotFoundException {
+        DentistaEntity dentista = dentistaRepository.findById(id).orElseThrow(
+                () -> new NotFoundException("Dentista não encontrado"));
         DentistaDTO dentistaDTO = mapperEntityToDTO(dentista);
         return dentistaDTO;
     }
 
-    public List<DentistaDTO> getByNome(String nome) {
+    public List<DentistaDTO> getByNome(String nome) throws NotFoundException {
         List<DentistaEntity> dentistas = dentistaRepository.getByNome(nome);
+        if (dentistas.isEmpty()) {
+            throw new NotFoundException("Dentista(s) de nome " + nome + " não encontrado(s)");
+        }
         return dentistas.stream().map(this::mapperEntityToDTO).toList();
     }
 
-    public DentistaDTO getByMatricula(String matricula) {
+    public DentistaDTO getByMatricula(String matricula) throws NotFoundException {
         DentistaEntity dentista = dentistaRepository.getByMatricula(matricula);
+        if (dentista == null) {
+            throw new NotFoundException("Dentista de matrícula " + matricula + " não encontrado");
+        }
         DentistaDTO dentistaDTO = mapperEntityToDTO(dentista);
         return dentistaDTO;
     }
 
     @Override
-    public List<DentistaDTO> getAll() {
+    public List<DentistaDTO> getAll() throws NotFoundException {
         List<DentistaEntity> dentistas = dentistaRepository.findAll();
+        if (dentistas.isEmpty()) {
+            throw new NotFoundException("Nenhum dentista cadastrado");
+        }
         return dentistas.stream().map(this::mapperEntityToDTO).toList();
     }
 
     @Override
-    public DentistaDTO update(DentistaDTO dentistaDTO, Long id) {
-        DentistaEntity dentistaEntity = dentistaRepository.findById(id).get();
+    public DentistaDTO update(DentistaDTO dentistaDTO, Long id) throws NotFoundException {
+        DentistaEntity dentistaEntity = dentistaRepository.findById(id).orElseThrow(
+                () -> new NotFoundException("Dentista não encontrado"));
         dentistaEntity.setNome(dentistaDTO.getNome());
         dentistaEntity.setSobrenome(dentistaDTO.getSobrenome());
         dentistaEntity.setUsername(dentistaDTO.getUsername());
@@ -75,9 +93,13 @@ public class DentistaServiceImpl implements IClinicaService<DentistaDTO>, UserDe
     }
 
     @Override
-    public String delete(Long id) {
+    public String delete(Long id) throws NotFoundException {
+        DentistaEntity dentista = dentistaRepository.findById(id).get();
+        if (dentista == null) {
+            throw new NotFoundException("Dentista não encontrado");
+        }
         dentistaRepository.deleteById(id);
-        return "Dentista excluído (ID: " + id + ")";
+        return "Dentista ID " + id + " excluido com sucesso";
     }
 
     private DentistaEntity mapperDTOToEntity(DentistaDTO dentistaDTO) {
@@ -98,19 +120,21 @@ public class DentistaServiceImpl implements IClinicaService<DentistaDTO>, UserDe
         return dentista;
     }
 
-    public DentistaEntity findById(Long dentista) {
-        return dentistaRepository.findById(dentista).orElseThrow(() -> {
-            throw new RuntimeException("Dentista não encontrado");
-        });
+    public DentistaEntity findById(Long dentista) throws NotFoundException {
+        DentistaEntity dentistaEntity = dentistaRepository.findById(dentista).orElseThrow(
+                () -> new NotFoundException("Dentista não encontrado"));
+        return dentistaEntity;
     }
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         if (pacienteRepository.findByUsername(username).isPresent()) {
-            return pacienteRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("userService: Paciente não encontrado"));
+            return pacienteRepository.findByUsername(username).orElseThrow(
+                    () -> new UsernameNotFoundException("Paciente não encontrado"));
         }
         if (dentistaRepository.findByUsername(username).isPresent()) {
-            return dentistaRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("userService: Dentista não encontrado"));
+            return dentistaRepository.findByUsername(username).orElseThrow(
+                    () -> new UsernameNotFoundException("Dentista não encontrado"));
         }
         return null;
     }
